text fix
