+++ frontend/kalunwa-web-app/src/app/admin/components/admin-home/admin-home.component.spec.ts
@@ -1,4 +1,8 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { RouterModule } from '@angular/router';
+import { RouterTestingModule } from '@angular/router/testing';
+import { routes } from 'src/app/app-routing.module';
 
 import { AdminHomeComponent } from './admin-home.component';
 
@@ -8,7 +12,8 @@ describe('AdminHomeComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ AdminHomeComponent ]
+      declarations: [ AdminHomeComponent ],
+      imports:[ReactiveFormsModule, FormsModule, RouterTestingModule.withRoutes(routes)]
     })
     .compileComponents();
   });
